Add files via upload
Assignment 2: L-Systems

I put in a lot of effort trying to understand L-systems, especially since I am working with algae in one of my projects. It was quite challenging, so I had to ask for some help along the way. :(

Here are some variations of L-systems that I’ve worked on.